wip: update useWsrx
