+++ src/GZCTF/ClientApp/src/components/InstanceEntry.tsx
@@ -8,9 +8,9 @@ import dayjs from 'dayjs'
 import duration from 'dayjs/plugin/duration'
 import { FC, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { HandleWsrxError } from '@Components/WsrxBox'
 import { getProxyUrl } from '@Utils/Shared'
 import { useConfig } from '@Hooks/useConfig'
+import { HandleWsrxError } from '@Hooks/useWsrx'
 import { useWsrx } from '@Hooks/useWsrx'
 import { ClientFlagContext } from '@Api'
 import misc from '@Styles/Misc.module.css'
