--- src/GZCTF/ClientApp/src/components/WsrxBox.tsx
@@ -1,129 +1,0 @@
-import { ActionIcon, Group, Stack, Text, TextInput } from '@mantine/core'
-import { showNotification } from '@mantine/notifications'
-import { mdiClose, mdiRefresh, mdiTuneVertical } from '@mdi/js'
-import Icon from '@mdi/react'
-import { WsrxError, WsrxErrorKind, WsrxState } from '@xdsec/wsrx'
-import { FC, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
-import { showErrorNotification } from '@Utils/ApiHelper'
-import { useConfig } from '@Hooks/useConfig'
-import { useWsrx } from '@Hooks/useWsrx'
-
-export const HandleWsrxError = (err: unknown, t: (key: string) => string) => {
-  if (err instanceof WsrxError) {
-    switch (err.kind) {
-      case WsrxErrorKind.VersionMismatch:
-        showNotification({
-          color: 'orange',
-          icon: <Icon path={mdiClose} size={1} />,
-          title: t('wsrx.error.version_mismatch.title'),
-          message: t('wsrx.error.version_mismatch.message'),
-        })
-        break
-
-      case WsrxErrorKind.DaemonUnavailable:
-        showNotification({
-          color: 'red',
-          icon: <Icon path={mdiClose} size={1} />,
-          title: t('wsrx.error.daemon_unavailable.title'),
-          message: t('wsrx.error.daemon_unavailable.message'),
-        })
-        break
-      case WsrxErrorKind.DaemonError:
-        showNotification({
-          color: 'red',
-          icon: <Icon path={mdiClose} size={1} />,
-          title: t('wsrx.error.daemon_error.title'),
-          message: t('wsrx.error.daemon_error.message'),
-        })
-        break
-      default:
-        showNotification({
-          color: 'red',
-          icon: <Icon path={mdiClose} size={1} />,
-          title: t('common.error.unknown'),
-          message: t('wsrx.error.unknown'),
-        })
-    }
-  } else {
-    showErrorNotification(err, t)
-  }
-}
-
-/**
- * WsrxBox component
- *
- * Wsrx's state will be managed by this component.
- */
-export const WsrxBox: FC<{}> = () => {
-  const { wsrx, wsrxOptions, setWsrxOptions } = useWsrx()
-  const platformConfig = useConfig()
-  const [wsrxState, setWsrxState] = useState(wsrx.getState())
-  const { t } = useTranslation()
-
-  wsrx.onStateChange((state) => {
-    setWsrxState(state)
-  })
-
-  const [showConfig, setShowConfig] = useState(false)
-
-  useEffect(() => {
-    if (platformConfig) {
-      setWsrxOptions({
-        ...wsrxOptions,
-        name: (platformConfig.config.title || 'GZ') + '::CTF',
-      })
-      wsrx.connect().catch(() => {})
-    }
-  }, [platformConfig.config.title])
-
-  const retryConnect = () => {
-    wsrx.connect().catch((err) => HandleWsrxError(err, t))
-  }
-
-  return (
-    <Stack gap="xs">
-      <Group h="30px" wrap="nowrap" justify="space-between" gap={2}>
-        {wsrxState === WsrxState.Usable && (
-          <Text fw="bold" c="green" flex={1}>
-            {t('wsrx.state.usable')}
-          </Text>
-        )}
-        {wsrxState === WsrxState.Pending && (
-          <Text fw="bold" c="orange" flex={1}>
-            {t('wsrx.state.pending')}
-          </Text>
-        )}
-        {wsrxState === WsrxState.Invalid && (
-          <Text fw="bold" c="orange" flex={1}>
-            {t('wsrx.state.invalid')}
-          </Text>
-        )}
-        <ActionIcon
-          variant="subtle"
-          aria-label={t('common.button.retry')}
-          onClick={retryConnect}
-          loading={wsrxState === WsrxState.Pending}
-        >
-          <Icon path={mdiRefresh} size={1} />
-        </ActionIcon>
-        <ActionIcon variant="subtle" aria-label={t('wsrx.button.config')} onClick={() => setShowConfig(!showConfig)}>
-          <Icon path={mdiTuneVertical} size={1} />
-        </ActionIcon>
-      </Group>
-      {showConfig && (
-        <>
-          <Group h="30px" wrap="nowrap" justify="space-between" gap={2}>
-            <TextInput
-              size="xs"
-              flex={1}
-              placeholder="127.0.0.1:3307"
-              value={wsrxOptions.api}
-              onChange={(e) => setWsrxOptions({ ...wsrxOptions, api: e.currentTarget.value })}
-            />
-          </Group>
-        </>
-      )}
-    </Stack>
-  )
-}
