+++ src/GZCTF/ClientApp/src/locales/en-US/wsrx.json
@@ -11,6 +11,10 @@
       "message": "The WSRX daemon is not running. If the desktop client has started, please check whether the WSRX control address in the platform settings is correct.",
       "title": "WSRX daemon unavailable"
     },
+    "daemon_offline": {
+      "message": "The WSRX daemon is offline. Please check if the WSRX daemon is running.",
+      "title": "WSRX daemon offline"
+    },
     "unknown": "An unknown error occurred while calling the WSRX API.",
     "version_mismatch": {
       "message": "WSRX version is too low, please update to the latest version before using platform integration.",
